feat: add profile & home
- add HomeSection Initial Structure
- add ProfileSection Initial Structure
- add profile & home page
- fix wagmi publicProvider
