+++ src/app/profile/page.tsx
@@ -1,0 +1,12 @@
+"use client";
+import { ProfileSection } from "@/components/04-templates";
+
+const Profile = () => {
+  return (
+    <div>
+      <ProfileSection />
+    </div>
+  );
+};
+
+export default Profile;
