+++ src/components/04-templates/index.tsx
@@ -1,2 +1,4 @@
 export * from "./404Section";
 export * from "./500Section";
+export * from "./ProfileSection";
+export * from "./HomeSection";
